use native types for DcMsg.viewtype (as done for most other constants, maintaining a separate type seems not to be worth the effort as it leads to confusion and bugs - alone 3 up to now: sticker is not image, videochat invitation was missing, type doubled at all)

+++ deltachat-ios/Chat/DraftModel.swift
@@ -22,7 +22,7 @@ public class DraftModel {
     }
     var viewType: Int32? {
         if let viewType = draftMsg?.type {
-            return Int32(viewType)
+            return viewType
         }
         return nil
     }
